Rename four rules for naming consistency
Renames:
- max-average-specificity → max-average-selector-specificity
- no-unknown-custom-property → no-unknown-custom-properties
- no-static-media-query → no-static-media-queries
- no-static-container-query → no-static-container-queries

Rationale:
- max-average-selector-specificity now matches the established
  max-average-selector-complexity / max-average-selectors-per-rule pattern
- no-unknown-custom-properties is now plural, consistent with every other
  rule in the no-unknown-* and no-unused-* family
- no-static-media-queries and no-static-container-queries are plural,
  consistent with max-unique-media-queries and no-unknown-container-names

Migration: find and replace in your stylelint config files:

  'projectwallace/max-average-specificity'
  → 'projectwallace/max-average-selector-specificity'

  'projectwallace/no-unknown-custom-property'
  → 'projectwallace/no-unknown-custom-properties'

  'projectwallace/no-static-media-query'
  → 'projectwallace/no-static-media-queries'

  'projectwallace/no-static-container-query'
  → 'projectwallace/no-static-container-queries'
