+++ src/configs/recommended.ts
@@ -20,10 +20,10 @@ export default {
 		'projectwallace/no-prefixed-values': true,
 		'projectwallace/no-property-browserhacks': true,
 		'projectwallace/no-property-shorthand': true,
-		'projectwallace/no-static-container-query': true,
-		'projectwallace/no-static-media-query': true,
+		'projectwallace/no-static-container-queries': true,
+		'projectwallace/no-static-media-queries': true,
 		'projectwallace/no-unknown-container-names': true,
-		'projectwallace/no-unknown-custom-property': true,
+		'projectwallace/no-unknown-custom-properties': true,
 		'projectwallace/no-unreachable-media-conditions': true,
 		'projectwallace/no-unused-container-names': true,
 		'projectwallace/no-unused-custom-properties': true,
@@ -60,7 +60,7 @@ export default {
 		'projectwallace/max-unique-units': 10,
 		'projectwallace/min-selector-uniqueness-ratio': 0.66,
 		'projectwallace/min-declaration-uniqueness-ratio': 0.5,
-		'projectwallace/max-average-specificity': [0.02, 2.5, 1],
+		'projectwallace/max-average-selector-specificity': [0.02, 2.5, 1],
 		'projectwallace/max-selectors-per-rule': 10,
 		'projectwallace/max-declarations-per-rule': 15,
 		'projectwallace/max-spacing-resets': 16,
