+++ src/index.test.ts
@@ -12,7 +12,7 @@ test('exports an array of stylelint rules', () => {
 		'projectwallace/max-average-declarations-per-rule',
 		'projectwallace/max-average-selector-complexity',
 		'projectwallace/max-average-selectors-per-rule',
-		'projectwallace/max-average-specificity',
+		'projectwallace/max-average-selector-specificity',
 		'projectwallace/max-comment-size',
 		'projectwallace/max-comments',
 		'projectwallace/max-declarations',
@@ -57,10 +57,10 @@ test('exports an array of stylelint rules', () => {
 		'projectwallace/no-prefixed-values',
 		'projectwallace/no-property-browserhacks',
 		'projectwallace/no-property-shorthand',
-		'projectwallace/no-static-container-query',
-		'projectwallace/no-static-media-query',
+		'projectwallace/no-static-container-queries',
+		'projectwallace/no-static-media-queries',
 		'projectwallace/no-unknown-container-names',
-		'projectwallace/no-unknown-custom-property',
+		'projectwallace/no-unknown-custom-properties',
 		'projectwallace/no-unreachable-media-conditions',
 		'projectwallace/no-unused-container-names',
 		'projectwallace/no-unused-custom-properties',
