+++ src/rules/no-static-media-queries/index.test.ts
@@ -2,7 +2,7 @@ import stylelint from 'stylelint'
 import { test, expect } from 'vitest'
 import plugin from './index.js'
 
-const rule_name = 'projectwallace/no-static-media-query'
+const rule_name = 'projectwallace/no-static-media-queries'
 
 const config = {
 	plugins: [plugin],
